server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "asyncpg>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]
